updated to 3.8
added random event
added tictactoe
chage to use streamelements poits

+++ src/main/java/me/alien/twitch/integration/util/ValueComparator.java
@@ -5,8 +5,8 @@ import java.util.Map;
 
 public class ValueComparator implements Comparator<String> {
 
-    Map<String, Integer> base;
-    public ValueComparator(Map<String, Integer> base) {
+    Map<String, Long> base;
+    public ValueComparator(Map<String, Long> base) {
         this.base = base;
     }
